+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Patched...
+
+### Added
+- Add an option `rocket-punch build --svg default`
+
+### Fixed
+- Exclude from result of `rocket-punch view` if can't get package metadata from NPM registry 
+
 ## [1.3.1] - 2020-07-20
 
 ### Added